requirements.txt
# -----------------------
# Web API
# -----------------------
fastapi>=0.110,<1.0
uvicorn[standard]>=0.29,<1.0
python-multipart>=0.0.9

# -----------------------
# Deep Learning / Inference
# -----------------------
ultralytics>=8.2.0
openvino>=2025.4.1,<2026.0
onnxruntime>=1.17.0

# -----------------------
# Tracking & Supervision
# -----------------------
supervision>=0.9.0

# -----------------------
# Vector Search / Embeddings
# -----------------------
faiss-cpu>=1.8.0

# -----------------------
# Computer Vision / Image Processing
# -----------------------
opencv-python-headless>=4.9.0
pillow>=10.0.0

# -----------------------
# Data Handling
# -----------------------
numpy>=1.26
pandas>=2.2.0

# -----------------------
# Dashboard / Visualizations
# -----------------------
streamlit>=1.32.0

pyyaml

# -----------------------
# Vector Search / Embeddings
# -----------------------
faiss-cpu>=1.8.0

# -----------------------
# Computer Vision / Image Processing
# -----------------------
opencv-python-headless>=4.9.0
pillow>=10.0.0

# -----------------------
# Data Handling
# -----------------------
numpy>=1.26
pandas>=2.2.0

# -----------------------
# Dashboard / Visualizations
# -----------------------
streamlit>=1.32.0

pyyaml

aiortc
av